refactor: move GTR inference counting to ancestral/, Deletion and resolve_indels to seq/
- Move `Deletion` type from `partition/payload/sparse` to `seq/indel` (pure gap count pair, not partition-specific)
- Move `resolve_indels_backward`, `resolve_indels_forward`, `IndelsBackward` from `ancestral/fitch_indel` to `seq/indel` (pure functions on gap ranges)
- Move `infer_gtr_fitch` and `get_mutation_counts_fitch` from `gtr/infer_gtr/fitch` to `ancestral/gtr_inference` (reads partition data, belongs with reconstruction pipeline)
- Move `infer_gtr_dense` and helpers from `gtr/infer_gtr/dense` to `ancestral/gtr_inference_dense` (dead code, retained for future use)
- Leave re-exports in `ancestral/fitch_indel` for backward compatibility within ancestral/ tests
- `gtr/infer_gtr/` retains only `common` (pure math solver) and `site_specific` (pure gtr/ deps)

Breaks two dependency cycles:
- `gtr/ -> partition/` eliminated (gtr/ no longer imports partition types)
- `partition/ -> ancestral/` reduced (marginal_dense now imports resolve_indels from seq/ instead of ancestral/)
