+++ packages/treetime/src/ancestral/__tests__/test_fitch_sub.rs
@@ -5,7 +5,8 @@ mod tests {
     finalize_sequence_forward, resolve_fixed_positions_backward, resolve_nonroot_substitutions_forward,
     resolve_root_forward, resolve_variable_positions_backward,
   };
-  use crate::partition::payload::sparse::{Deletion, FitchSeqDistribution, SparseEdgePartition, SparseSeqInfo};
+  use crate::partition::payload::sparse::{FitchSeqDistribution, SparseEdgePartition, SparseSeqInfo};
+  use crate::seq::indel::Deletion;
   use crate::seq::composition::Composition;
   use eyre::Report;
   use maplit::btreemap;
