+++ packages/treetime/src/ancestral/fitch_indel.rs
@@ -1,120 +1,1 @@
-use crate::partition::payload::sparse::Deletion;
-use crate::seq::indel::InDel;
-use std::collections::BTreeMap;
-use treetime_primitives::Seq;
-use treetime_utils::interval::range_complement::range_complement;
-use treetime_utils::interval::range_difference::range_difference;
-use treetime_utils::interval::range_intersection::range_intersection;
-
-pub struct IndelsBackward {
-  pub variable_indel: BTreeMap<(usize, usize), Deletion>,
-  pub resolved_gaps: Vec<(usize, usize)>,
-}
-
-/// Resolve indels during the backward pass: identify positions where children
-/// disagree on gap presence.
-///
-/// Returns variable indels (unresolved disagreements) and resolved gaps (ranges
-/// where all children agree on gap).
-pub fn resolve_indels_backward(
-  child_gaps: &[Vec<(usize, usize)>],
-  child_variable_indels: &[&BTreeMap<(usize, usize), Deletion>],
-  length: usize,
-) -> IndelsBackward {
-  let n_children = child_gaps.len();
-  let consensus_gaps = range_intersection(child_gaps);
-  let non_gap = range_complement(&[(0, length)], &[consensus_gaps]);
-
-  let mut variable_indel = BTreeMap::new();
-
-  for child_gap in child_gaps {
-    for r in range_intersection(&[non_gap.clone(), child_gap.clone()]) {
-      let indel = variable_indel.entry(r).or_insert_with(|| Deletion {
-        deleted: 0,
-        present: n_children,
-      });
-      indel.deleted += 1;
-      indel.present -= 1;
-    }
-  }
-
-  for child_vi in child_variable_indels {
-    for r in child_vi.keys() {
-      if let Some(indel) = variable_indel.get_mut(r) {
-        indel.deleted += 1;
-        indel.present -= 1;
-      }
-    }
-  }
-
-  let mut resolved_gaps = Vec::new();
-  variable_indel.retain(|r, indel| {
-    if indel.deleted == n_children {
-      resolved_gaps.push(*r);
-      false
-    } else {
-      true
-    }
-  });
-
-  IndelsBackward {
-    variable_indel,
-    resolved_gaps,
-  }
-}
-
-/// Resolve variable indels during the forward pass using parent context.
-///
-/// For each variable indel range, uses parent gap state to break ties.
-/// Also detects consensus gap differences between node and parent (non-variable
-/// deletions/insertions).
-///
-/// Returns the list of resolved InDel entries for the edge.
-///
-/// Dense partitions do not need `InDel::seq` content for inserted sequences
-/// because the full sequence is available from the dense reconstruction.
-/// We populate it for consistency with sparse indels.
-pub fn resolve_indels_forward(
-  variable_indel: &BTreeMap<(usize, usize), Deletion>,
-  node_gaps: &mut Vec<(usize, usize)>,
-  parent_gaps: &[(usize, usize)],
-  parent_sequence: &Seq,
-  node_sequence: &Seq,
-) -> Vec<InDel> {
-  let mut indels = Vec::new();
-
-  // Process variable indels using parent context for tiebreaking
-  for (r, indel) in variable_indel {
-    let gap_in_parent = if parent_gaps.contains(r) { 1 } else { 0 };
-    if indel.deleted + gap_in_parent > indel.present {
-      node_gaps.push(*r);
-      if gap_in_parent == 0 {
-        let indel = InDel::del(*r, &parent_sequence[r.0..r.1]);
-        indels.push(indel);
-      }
-    } else if gap_in_parent > 0 {
-      let indel = InDel::ins(*r, &node_sequence[r.0..r.1]);
-      indels.push(indel);
-    }
-  }
-
-  // Process consensus gaps in node that are not in parent (deletions)
-  for r in range_difference(node_gaps, parent_gaps) {
-    if variable_indel.contains_key(&r) {
-      continue;
-    }
-    let indel = InDel::del(r, &node_sequence[r.0..r.1]);
-    indels.push(indel);
-  }
-
-  // Process gaps in parent that are not in node (insertions)
-  for r in range_difference(parent_gaps, node_gaps) {
-    if variable_indel.contains_key(&r) {
-      continue;
-    }
-    let indel = InDel::ins(r, &node_sequence[r.0..r.1]);
-    indels.push(indel);
-  }
-
-  indels
-}
+pub use crate::seq::indel::{Deletion, IndelsBackward, resolve_indels_backward, resolve_indels_forward};
