+++ packages/treetime/src/gtr/infer_gtr/__tests__/test_gm_infer_gtr_dense.rs
@@ -10,7 +10,7 @@ mod tests {
   use crate::ancestral::marginal::initialize_marginal;
   use crate::gtr::get_gtr::{JC69Params, jc69};
   use crate::gtr::infer_gtr::common::InferGtrResult;
-  use crate::gtr::infer_gtr::dense::infer_gtr_dense;
+  use crate::ancestral::gtr_inference_dense::infer_gtr_dense;
   use crate::pretty_assert_ulps_eq;
   use crate::partition::marginal_dense::PartitionMarginalDense;
   use crate::partition::payload::ancestral::GraphAncestral;
