+++ packages/treetime/src/gtr/infer_gtr/mod.rs
@@ -1,6 +1,4 @@
 #[cfg(test)]
 mod __tests__;
 pub(crate) mod common;
-pub(crate) mod dense;
-pub(crate) mod fitch;
 pub mod site_specific;
