+++ packages/treetime/src/partition/payload/dense.rs
@@ -1,5 +1,5 @@
 use crate::alphabet::alphabet::Alphabet;
-use crate::partition::payload::sparse::Deletion;
+use crate::seq::indel::Deletion;
 use crate::seq::find_char_ranges::find_letter_ranges;
 use crate::seq::indel::InDel;
 use eyre::Report;
